Move to CreatePoint tab when tapping FAB in Main
This will be a screen used to create new points and send it to the
server. Right now there's nothing there. Also there's no button to go
back on that screen so you have to press the back button.

It's mostly there right now as an example of how to move to a new
Activity.